requirements.txt
snowflake-sqlalchemy
rudder-sdk-python
scikit-learn
boto3
seaborn
matplotlib
pandas
numpy
pyyaml
tqdm
ipython
sagemaker
jupyter
fastparquet
pyarrow
MarkupSafe
papermill

